keymap: When flattening the map, use the first value
We have a few cases where we have special names for a combination of keys that
are also present in the "modified" tables. We want the special names. As we
always put the non-modified stuff in first, if we never overwrite values in the
flattened table, we're good. So lets do that.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>