keymap: Don't include Hyper & Meh in modifiers
We want to use the names as short codes, but at the same time, we want to show
the real keys underneath them when working with them. For this reason, don't
include them in `ModifiersTable`, but in `HyperMehTable`, which gets added
later, and isn't part of the basic set (thus, it is used for display, but not
for editing).

This also means that the table flattening logic had to be reverted, because
we *do* want overrides now.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>