App.js: Fix onKeyboardConnect()'s navigate targets
We should be navigating to `/layout-editor` or `/welcome`, not `./layout-editor`
or `./welcome`. The distinction is important, the router doesn't know how to
handle the latter ones.

Fixing this navigation means we do not need the one in KeyboardSelect anymore.

These two together restore the functionality of landing on the Welcome screen
when Focus wasn't detected on the keyboard.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>