Merge and redesign the layout & colormap editors
Instead of having a separate layout- and colormap editor, provide both on the
same screen. To achieve this, we needed to compact the secondary toolbar, thus
the layer tabs were turned into a dropdown instead. Chosing between keymap and
colormap editing can be done with a new pair of toggle buttons. The colormap
stuff is not offered if the board does not support colormap.

While there, the "..." (more) menu on the right side of the secondary toolbar
has been unpacked, the copy & clear functions were lifted to the main toolbar,
as icons. Clearing now prompts for a confirmation (fixes #316), and layer
copying pops up a dialog with an expanded layer list to choose from.

With the merge, we now display keys on top of the colormap, thus fixes #303 and
also fixes #258, because we do not have separate screens anymore. This also
fixes #272, for the same reason.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>