focus: Support more reliable keyboard detection during enumeration
During enumeration, we can now call `focus.isDeviceSupported(port)`, which will
dispatch to the hardware plugin, to see if this _variant_ of the board is
supported or not. The implementation defaults to marking devices supported,
unless the hardware plugin decides otherwise.

The ANSI and ISO variants of Dygma's Raise were adjusted to provide this new
method, and allow only the appropriate variant through.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>
Signed-off-by: Roman Danilov <r.danylov@gmail.com>