Editor: Support cases where we have a colormap only
The Editor screen already had support for cases where we only have a keymap, but
no colormap; it also supported having both; but it lacked support for the case
where we only have a colormap.

There are many valid reasons for not having an editable keymap, and not having
the `EEPROMKeymap` plugin enabled at all. This patch reworks the Editor screen a
bit so that we support firmware with said plugin disabled, too.

Fixes #496.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>