Add a progress bar to the FirmwareUpdate screen
Knowing what's happening, what we're waiting for, is very valuable information,
both for the user, and for developers too, when working on bug reports. To this
end, lets display a progress bar when flashing is in progress, giving a rough
outline about where we are in the process.

At the moment, this is only implemented for boards using the Avr109 flashing
protocol. The others will display a progress stepper too, but with a single
element only, for now.

Fixes #492.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>