Optionally disable the sandbox on the AppImage build
On Linux, sandboxing requires either a suid root helper (which we can't ship in
the AppImage build), running as root (which is something we'd rather avoid), or
unprivileged user namespace cloning (which is only available on more recent
kernels). Since we do not want to run as root, and we'd like users to run
Chrysalis without having to explicitly pass arguments to it, on Linux, we wrap
the launcher.

The wrapper checks for the necessary kernel feature, and disables the sandbox if
the feature is not available. This is still a better and safer experience than
running as root, or than the user having to explicitly run Chrysalis with
`--no-sandbox`.

This is only done on Linux, and only for the AppImage build. Running `yarn run
start` will still require the extra flag.

Fixes #499.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>