Merge the Bootloader device descriptors into the main one
We want to be able to know which bootloader descriptor corresponds to which
device, so that when we reset one during flashing, we can find the appropriate
bootloader device. For this reason, the *Bootloader descriptors were merged into
the main device descriptors, and the code was changed to account for this.

In practice, this doesn't change much. The only visible change at the moment is
that a device in bootloader mode will show the device path on the key selector
screen now, and will look the same as a normal keyboard. When connecting to it,
we'll be taken to the screen that prompts us to upgrade the firmware.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>