Major redesign of the Editor screen
Our Editor screen has been showing its age, and with this massive change, we
replace it entirely. We do so in one big commit, because it is a major rewrite,
splitting it up into smaller parts is not practical.

With the new Editor, we moved away from having a double toolbar at the top, and
a selector on the bottom: both were merged into a sidebar, with a number of
sections for keys and features. Many of those will show their contents in-line,
some of them, like the key picker, will pop up a dialog instead.

We also redesigned how our labels are displayed, and we now support different
host-side layouts, via the Unicode CLDR project. While there's work to be done
on the labeling front, the APIs to do so, and to work with keys and key groups,
are considerably more developer friendly.

Furthermore, layout sharing has been redesigned aswell: we now export and import
to and from files, and share the whole keymap, all layers included. At this
time, there's no per-layer sharing supported.

As a regression, layer copying and layer erasing is currently not implemented
for the new Editor.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>