Editor: Group the host Layouts by language
Using the `cldr` package (not ours, a 3rd party one), we can turn a language
code into a display name. Using this, and the new `group` key in the layout
database, we can display the available layouts grouped by language.

We're using a 3rd party library, a non-trivial one at that, because it supports
translating the language code into different display languages too, so we will
not have to do that ourselves, once we start caring about languages other than
English.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>