Fix the installation of the udev rules
When installing the udev rules from the AppImage, we need to copy the rules to a
temporary place first, before elevating privileges, and then copy from there to
the final place. We need to do this because the AppImage temporary mount is
locked down, and not even root can access it, so we need to get the source file
out of there first.

Fixes #603.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>