github/workflows: Adjust the snapshot/prod publish conditions
We moved away from tagging production releases with a `chrysalis-` prefix, we're
using `-snapshot` postfixes for snapshot builds instead.

In the end, we'll publish a snapshot for every push to master (even if it's a
release push), and we'll publish a release if we're tagged appropriately.

This addresses the second part of #644.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>