Editor: Move the warning about mixed layers to its own screen
When encountering a keyboard with mixed hardcoded and eeprom layers - which we
no longer support in Chrysalis - instead of displaying an alert below the
toolbar and otherwise letting the user proceed and get confused, show a separate
screen explaining the situation, and offering to switch to custom layers only.

The new screen still allows opening a feature request, but being a separate
screen, redirected to from the Editor, it no longer allows editing the keymap.

Additionally, when switching to custom layers only, we also set the default
layer to zero, lest we end up in a situation where the default layer is out of
bounds.

This addresses #658.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>