Update our release and snapshot workflows
This is a major restructuring of our GitHub Actions workflows. First of all, we
now have separate workflows for release and snapshot builds, as having them
separate makes it easier to control which one runs when. It also allows us to
fine tune a lot of things for the snapshot workflow, which runs much more often.

Another change is that we're now using a static tag (`snapshot`) for development
builds, rather than a versioned tag, and this tag - and the pre-release on
GitHub - will get remade on every run, to keep it up to date.

We also have the build number back for our snapshot builds again, something we
lost with the transition away from Travis.

Documentation and scripts have been appropriately updated.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>