api/keymap: Make it possible to migrate legacy keycodes to new ones
From time to time, we may end up having to migrate keycodes from one set to
another, as it happened with Consumer keys. To make the migration process
easier, we want to store the old, legacy codes along the new ones, and have a
way to migrate from legacy to current codes.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>