Improve legacy keycode lookup & migration
Instead of iterating over every code in the database to find legacy codes, add
them to the flattened array with a `legacy: true` flag. This speeds up lookups
considerably. Additionally, try to scan the keymap for legacy codes only when
necessary, not on every single render of the Editor screen.

These two drastically speed up the key changing process, which suffered a big
slowdown with the introduction of legacy keycodes.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>